# [CMake] Speed up copying of headers