--- a/CMakeLists.txt
+++ b/CMakeLists.txt
@@ -388,8 +388,7 @@ ROOT_ADD_TEST_SUBDIRECTORY(tutorials)
 ROOT_CREATE_HEADER_COPY_TARGETS()
 
 get_property(__allHeaders GLOBAL PROPERTY ROOT_HEADER_TARGETS)
-get_property(__allBuiltins GLOBAL PROPERTY ROOT_BUILTIN_TARGETS)
-add_custom_target(move_headers ALL DEPENDS ${__allHeaders} ${__allBuiltins} gitinfotxt)
+add_custom_target(move_headers DEPENDS ${__allHeaders} gitinfotxt)
 
 #---CXX MODULES-----------------------------------------------------------------------------------
 if(MSVC)