--- a/builtins/xxhash/CMakeLists.txt
+++ b/builtins/xxhash/CMakeLists.txt
@@ -30,5 +30,3 @@ add_library(xxHash::xxHash ALIAS xxhash)
 
 set(xxHash_LIBRARY $<TARGET_FILE:xxhash> CACHE INTERNAL "")
 set(xxHash_LIBRARIES xxHash::xxHash CACHE INTERNAL "")
-
-set_property(GLOBAL APPEND PROPERTY ROOT_BUILTIN_TARGETS xxHash::xxHash)