… builtins/glew/CMakeLists.txt
… builtins/lz4/CMakeLists.txt
… builtins/xrootd/CMakeLists.txt
… builtins/zlib/CMakeLists.txt
… builtins/zstd/CMakeLists.txt