- ~~Move the copy of artifacts (e.g. tutorials) from build time to configure time, and use a faster copy function.~~
- At build time, replace the one-by-one copy of headers by copy in batches.

Here are some results:
| `make -j6 move_headers`  | New | Old |
|--------|--------|--------|
| 1st invocation | 10.7 s | 34.5 s |
| 2nd invocation | 2.7 s | 3.8 s |
| Invocation after touching a header in core/ | 3.1 s | 4.2 s | 
| Invocation after touching a header in RooFit/hs3 | 2.9 s | 3.9 s |

For the copy of artifacts, there is no agreement yet since it's more complicated. I will factor this part out into a second PR.

Fix #14953