@hageboeck 2025-02-05
> The error on Windows is related to this PR:
> 
> ```
> error MSB8066: Custom build for 'C:\ROOT-CI\build\CMakeFiles\b4bf4ec5a2a2748fa3fe14aac0b491c8\Bswapcpy.h.rule;C:\ROOT-CI\build\CMakeFiles\d52c4a97b544285cfdcb9662e36a1605\BasicFCN.h.rule;C:\ROOT-CI\build\CMakeFiles\5cbe746a32c517ce8263460eaed96b46\eglew.h.rule;C:\ROOT-CI\build\CMakeFiles\6f62a2407de5b487a7d37393f911fe87\AdaptiveIntegratorMultiDim.h.rule;C:\ROOT-CI\build\CMakeFiles\f946b1f1f35c1bd8f320cb3e49e00bca\3DConversions.h.rule;C:\ROOT-CI\build\CMakeFiles\d4bf7379258006facfcfa53f75f5956b\ABObj.h.rule;C:\ROOT-CI\build\CMakeFiles\a35c8b97151be13ed3d204ad6ed80264\RAnyObjectHolder.hxx.rule;C:\ROOT-CI\build\CMakeFiles\456e85c62d06e0cd87d98ee60dcc2e4f\EExecutionPolicy.hxx.rule;C:\ROOT-CI\build\CMakeFiles\c353aed7a3bd6dbf0923bb9dc4fba686\ActionHelpers.hxx.rule;C:\ROOT-CI\build\CMakeFiles\9bbd6f9b9a1ccee3bcd008c06562308d\RFieldFundamental.hxx.rule;C:\ROOT-CI\build\CMakeFiles\3fbab4c4bb0860c35dbc49b686024422\daos.h.rule;C:\ROOT-CI\build\CMakeFiles\06709d940c4eaf4b90467c901d21fad5\CodegenImpl.h.rule;C:\ROOT-CI\build\CMakeFiles\2600cfd80747fb5960de4549ed2956e4\MathFuncs.h.rule;C:\ROOT-CI\build\CMakeFiles\193c5d220ee00c87398efe41e5bca113\LikelihoodGradientWrapper.h.rule;C:\ROOT-CI\build\CMakeFiles\7e106373c3f52de7795b17f694a854fe\JSONIO.h.rule;C:\ROOT-CI\build\CMakeFiles\c562cae4eae5394921ddc3d0cafa5f56\RooCatTypeLegacy.h.rule;C:\ROOT-CI\build\CMakeFiles\ebec96819e385d9ea45a1b2d103e31b8\AsymptoticCalculator.h.rule;C:\ROOT-CI\build\CMakeFiles\cd2d88b30ee96ff2fdca17cad3e08856\Asimov.h.rule;C:\ROOT-CI\build\CMakeFiles\0bd4f58f57ed53aeade166f8902ab30f\HistFactoryImpl.h.rule;C:\ROOT-CI\build\CMakeFiles\4ff935d273915365070a61ee0cd1b02c\BDTEventWrapper.h.rule;C:\ROOT-CI\build\CMakeFiles\db29aecca489daaa0f06f6367b1d3b81\RBatchGenerator.hxx.rule;C:\ROOT-CI\build\CMakeFiles\ec821efde375e08e4c351de8c63c89b6\Adadelta.h.rule;C:\ROOT-CI\build\CMakeFiles\addb20638c349f1661791f2822729cf7\Cpu.h.rule;C:\ROOT-CI\build\CMakeFiles\1690471466c052fbb378442159d120a1\CpuBuffer.h.rule;C:\ROOT-CI\build\CMakeFiles\c9afbfdea5ae89c0c83ece15ff85d3eb\DataLoader.h.rule;C:\ROOT-CI\build\CMakeFiles\c0ea46944f6274260ea7991effa94d3e\ContextHandles.h.rule;C:\ROOT-CI\build\CMakeFiles\36bb35dee450defb44862eb7f6abee62\GRULayer.h.rule;C:\ROOT-CI\build\CMakeFiles\d580fcda3530414ad56cc58ed53c3de0\bbox.h.rule;C:\ROOT-CI\build\CMakeFiles\c9dbea440c663f8cf1755fecccca54fd\TF1Data.h.rule;C:\ROOT-CI\build\CMakeFiles\837d695c63a67b677c4a78432d56e457\move_headers.rule;C:\ROOT-CI\src\CMakeLists.txt' exited with code 9020. [C:\ROOT-CI\build\move_headers.vcxproj]
> ```

@bellenot I might have found the problem: The command line seems to be too long. After splitting the copy command in packs of 100 (there is more than 1000 headers in otherwise), the build seemed to succeed. Let's wait for the latest run.