@hageboeck 2025-02-06
> With the copy of the artefact (seemingly) only at configure time, we have the following 'failures':
> 
> ```
> $ echo >> src/tutorials/hsimple.C 
> $ ls -l src/tutorials/hsimple.C tutorials/hsimple.C 
> -rw-r--r--. 1 pcanal us_cms 3572 Feb  5 17:41 src/tutorials/hsimple.C
> -rw-r--r--. 1 pcanal us_cms 3571 Feb  5 16:23 tutorials/hsimple.C
> $ ninja
> [2/2] Updating etc/gitinfo.txt.
> -- Found Git: /usr/bin/git (found version "2.43.5") 
> $ ls -l src/tutorials/hsimple.C tutorials/hsimple.C 
> -rw-r--r--. 1 pcanal us_cms 3572 Feb  5 17:41 src/tutorials/hsimple.C
> -rw-r--r--. 1 pcanal us_cms 3571 Feb  5 16:23 tutorials/hsimple.C
> ```
> 
> I.e. the copied `hsimple.C` is not updated after a local modification in the source.

That was my intention. I wanted to reduce the number of CMake targets, since I consider the macros and tutorials mostly stable.
I would anyway recommend to `root.exe <sourceDir>/<macro>` instead of `make -jX; root.exe <buildDir>/<macro>`