@hageboeck 2025-02-10
@pcanal I decided to factor out the copy of artifacts into a second PR, so at least the improvement due to the faster copying of headers can be merged already. Could you have a look again?