@pcanal 2025-02-10
> I decided to factor out the copy of artifacts into a second PR

Great.  I was about to propose that :)